backend/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backcall==0.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==2.2.16
django-colorfield==0.7.2
django-cors-headers==3.11.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.0.2
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flake8==5.0.4
idna==3.4
importlib-metadata==1.7.0
ipython-genutils==0.2.0
itypes==1.2.0
Jinja2==3.1.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
MarkupSafe==2.1.1
mccabe==0.7.0
mistune==2.0.4
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
zipp==3.9.0
